+++ Sources/TunnelKitOpenVPNCore/Configuration.swift
@@ -205,9 +205,6 @@ extension OpenVPN {
         /// The optional TLS wrapping.
         public var tlsWrap: TLSWrap?
         
-        /// If set, overrides TLS security level (0 = lowest).
-        public var tlsSecurityLevel: Int?
-        
         /// Sends periodical keep-alive packets if set.
         public var keepAliveInterval: TimeInterval?
         
@@ -324,7 +321,6 @@ extension OpenVPN {
                 clientCertificate: clientCertificate,
                 clientKey: clientKey,
                 tlsWrap: tlsWrap,
-                tlsSecurityLevel: tlsSecurityLevel,
                 keepAliveInterval: keepAliveInterval,
                 keepAliveTimeout: keepAliveTimeout,
                 renegotiatesAfter: renegotiatesAfter,
@@ -406,9 +402,6 @@ extension OpenVPN {
         
         /// - Seealso: `ConfigurationBuilder.tlsWrap`
         public let tlsWrap: TLSWrap?
-
-        /// - Seealso: `ConfigurationBuilder.tlsSecurityLevel`
-        public let tlsSecurityLevel: Int?
 
         /// - Seealso: `ConfigurationBuilder.keepAliveInterval`
         public let keepAliveInterval: TimeInterval?
@@ -527,7 +520,6 @@ extension OpenVPN.Configuration {
         builder.clientCertificate = clientCertificate
         builder.clientKey = clientKey
         builder.tlsWrap = tlsWrap
-        builder.tlsSecurityLevel = tlsSecurityLevel
         builder.keepAliveInterval = keepAliveInterval
         builder.keepAliveTimeout = keepAliveTimeout
         builder.renegotiatesAfter = renegotiatesAfter
@@ -585,11 +577,6 @@ extension OpenVPN.Configuration {
             log.info("\tTLS wrapping: \(tlsWrap.strategy)")
         } else {
             log.info("\tTLS wrapping: disabled")
-        }
-        if let tlsSecurityLevel = tlsSecurityLevel {
-            log.info("\tTLS security level: \(tlsSecurityLevel)")
-        } else {
-            log.info("\tTLS security level: default")
         }
         if let keepAliveSeconds = keepAliveInterval, keepAliveSeconds > 0 {
             log.info("\tKeep-alive interval: \(keepAliveSeconds.asTimeString)")